Implemented Feature : Facility to upload profile photos and to update it.

+++ saveListing.php
@@ -51,9 +51,9 @@ if($_SERVER['REQUEST_METHOD']=="POST"){
 				$fileToUpload = $_FILES["fileToUpload"]["name"];
 				$target_file = $target_dir.basename($fileToUpload);
 				$uploadOk = 1;
+				$alreadyExists = false;
 				$imageFileType = strtolower(pathinfo($target_file,PATHINFO_EXTENSION));
 
-				// Check if image file is a actual image or fake image
 				if(isset($_POST["submit"])) {
 				    $check = getimagesize($_FILES["fileToUpload"]["tmp_name"]);
 				    if($check !== false) {
@@ -65,33 +65,34 @@ if($_SERVER['REQUEST_METHOD']=="POST"){
 				    }
 				}
 
-				// Check if file already exists
 				if (file_exists($target_file)) {
-				    $_SESSION['message']="Sorry, file already exists.";
-				    $uploadOk = 0;
+				    $alreadyExists = true;
 				}
 
-				// Check file size
 				if ($_FILES["fileToUpload"]["size"] > 500000) {
 				    $_SESSION['message']="Sorry, your file is too large.";
 				    $uploadOk = 0;
 				}
 
-				// Allow certain file formats
 				if($imageFileType != "jpg" && $imageFileType != "png" && $imageFileType != "jpeg"
 				&& $imageFileType != "gif" ) {
 				    $_SESSION['message']="Sorry, only JPG, JPEG, PNG & GIF files are allowed.";
 				    $uploadOk = 0;
 				}
 
-				// Check if $uploadOk is set to 0 by an error
 				if ($uploadOk == 0) {
 				    $_SESSION['message']="Sorry, your file was not uploaded.";
-				// if everything is ok, try to upload file
 				} 
 
 				else{
-				    if (move_uploaded_file($_FILES["fileToUpload"]["tmp_name"], $target_file)) {
+					if(!$alreadyExists){
+						if(move_uploaded_file($_FILES["fileToUpload"]["tmp_name"], $target_file)){
+						}
+						else {
+					        echo "Sorry, there was an error uploading your file.";
+					    }
+					}
+				    else{
 				        $_SESSION['message']="The file ". basename( $_FILES["fileToUpload"]["name"]). " has been uploaded.";
 				        $imagePath = $conn->real_escape_string('uploads/'.$_FILES['fileToUpload']['name'].' ');
